changed matrix and vector types to encapsulate a double array instead of an int array

+++ Matrix.cs
@@ -6,20 +6,20 @@
         //I created this class because, while at it's core it is little more than an 2d integer array (int[,]),
         //I wanted the freedom to use these arrays in particular ways and be able to name them Matrices for clarity 
         //</summary>
-        private int[,] matrix;
+        private double[,] matrix;
         private int rows;
         private int columns;
         public Matrix(int xDimension, int yDimension)//Creating an empty matrix of (m, n) size
         {
-            this.matrix = new int[xDimension, yDimension];
+            this.matrix = new double[xDimension, yDimension];
             this.rows = this.GetRows();//Accesses the internal "GetRows()" method for the private matrix array
             this.columns = this.GetColumns();//Same as above but for the columns
         }
-        public Matrix(int[,] entries)//Creating a matrix given an array of integers
+        public Matrix(double[,] entries)//Creating a matrix given an array of integers
         {
             this.rows = this.GetRows(entries);//This is the method for the external getRows of something unknown to the object instance
             this.columns = this.GetColumns(entries);//Same as above but for the columns
-            this.matrix = new int[this.rows, this.columns];
+            this.matrix = new double[this.rows, this.columns];
             for (int i = 0; i < this.rows; i++)
             {
                 for (int j = 0; j < this.columns; j++)
@@ -32,7 +32,7 @@
         {
 
         }
-        public int this[int Xindex, int Yindex]
+        public double this[int Xindex, int Yindex]
         {
             get => this.matrix[Xindex, Yindex];//Not in standard { get; set; } form b/c I might alter index protection
             set => this.matrix[Xindex, Yindex] = value;
@@ -51,7 +51,7 @@
         }
         public Vector DotProduct(Vector inVector)
         {
-            int sigma = 0; //Used in finding summation (sigma) of each row
+            double sigma = 0; //Used in finding summation (sigma) of each row
 
             Vector final = new Vector(this.rows);//Create a vector in which you can store the values
 
@@ -73,11 +73,11 @@
         }
         public int GetColumns() => this.matrix.Length / this.matrix.GetLength(0); //Divid the total amt of elements by a column length to get row length
         public int GetRows() => this.matrix.GetLength(0);
-        public int GetColumns(int[,] input)//same as get columns but used for input
+        public int GetColumns(double[,] input)//same as get columns but used for input
         {
             return input.Length / input.GetLength(0);
         }
-        public int GetRows(int[,] input)
+        public int GetRows(double[,] input)
         {
             return input.GetLength(0);
         }
